requirements.txt
pytest-playwright
playwright
discord.py
python-dotenv
PyYAML
pytest